refactor: replace as-casts with Reg::from and 0u32.. iterators
Replace all narrowing `as Reg` / `as u8` casts in register arithmetic
with type-safe alternatives:

- u8 → Reg: use Reg::from(n) (infallible, compiler-verified)
- loop indices: use `for i in 0..Reg::from(n)` or `(0u32..).zip(iter)`
  so `i` is already Reg — no cast at the use site
- local_index stored as HashMap<String, Reg> (was usize) in all crates
  to avoid the cast in local_reg(); only widening `slot as usize` remains
  when indexing into Vec<Option<Value>> (u32 ≤ usize, always lossless)

+++ crates/pyrust-core/src/lib.rs
@@ -115,7 +115,7 @@ pub struct UserFunction {
     pub name: String,
     pub params: Vec<UserFunctionParam>,
     pub local_names: NameSet,
-    pub local_index: Rc<HashMap<String, usize>>,
+    pub local_index: Rc<HashMap<String, u32>>,
     pub global_names: NameSet,
     pub nonlocal_names: NameSet,
     pub env: EnvRef,
